Fix trailing comments in semicolon checking and logic error in Binary Search snippet

+++ js/editor/snippets.js
@@ -6,7 +6,7 @@ int binarySearch(int[] arr, int target) {
     int left = 0;
     int right = arr.length - 1;
     while (left <= right) {
-        int mid = left + (right - left) / 2;
+        int mid = (int) (left + (right - left) / 2);
         if (arr[mid] == target) {
             return mid;
         }
